pyproject.toml
[tool.poetry]
name = "ai-tracker"
version = "0.1.0"
description = "Telegram productivity bot"
authors = ["Nikita <example@example.com>"]
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.13.1"
tortoise-orm = "^0.21.7"
redis = "^5.0.8"
python-dotenv = "^1.0.1"
httpx = "^0.27.2"
asyncpg = "^0.29.0"

[tool.poetry.group.dev.dependencies]

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"


